project1/code/requirements.txt
bert-tensorflow
tensorflow
tensorflow-hub
nltk
sklearn
pandas